Use `Int64` in `CostModel`
See: https://github.com/IntersectMBO/cardano-ledger/commit/7c59e3a23033f8e4fdb4bf5b03b224e2c3752ffd

+++ cardano-api/internal/Cardano/Api/ProtocolParameters.hs
@@ -145,6 +145,7 @@ import           Data.Bifunctor (bimap, first)
 import           Data.ByteString (ByteString)
 import           Data.Data (Data)
 import           Data.Either.Combinators (maybeToRight)
+import           Data.Int (Int64)
 import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import           Data.Maybe (isJust)
@@ -1082,7 +1083,7 @@ fromAlonzoPrices Alonzo.Prices{Alonzo.prSteps, Alonzo.prMem} =
 -- Script cost models
 --
 
-newtype CostModel = CostModel [Integer]
+newtype CostModel = CostModel [Int64]
   deriving (Eq, Show, Data)
   deriving newtype (ToCBOR, FromCBOR)
